Move functions to pyanaconda.ui.lib.storage
Move functions from pyanaconda.storage.utils that are used only by UI
to pyanaconda.ui.lib.storage.
